+++ R/mesh_plot.R
@@ -23,8 +23,7 @@
 #' define the target map projection in 'PROJ string' format. (There is no
 #' "reproject" function for quadmesh, this is performed directly on the x-y
 #' coordinates of the 'quadmesh' output). The 'col' argument are mapped to the
-#' inputdata as in [graphics::image()], and applied relative to 'zlim' if
-#' also supplied.
+#' inputdata as in [graphics::image()].
 #'
 #' The `coords` argument only applies to a raster object. The `crs` argument
 #' only applies to a spatial object that has a crs projection metadata string
@@ -43,13 +42,13 @@
 #'
 #' If the input is a mesh3d and has a material texture image this is _approximated_
 #' by averaging the RGB values of each primitive's corner into a constant colour
-#' for that face.
+#' for that face. If you would like to avoid this texture colour, set the
+#' 'mesh3d$material$color' property to `NULL`.
 #' @param x object to convert to mesh and plot
 #' @param crs target map projection
 #' @param col colours to use, defaults to that used by [graphics::image()]
 #' @param add add to existing plot or start a new one
-#' @param zlim absolute range of data to use for colour scaling (if `NULL` the
-#'   data range is used)
+#' @param zlim unimplemented (was used in 'quadmesh::mesh_plot')
 #' @param ... passed through to `base::plot`
 #' @param coords optional input raster of coordinates of each cell, see details
 #' @return nothing, used for the side-effect of creating or adding to a plot
@@ -72,30 +71,21 @@ mesh_plot.mesh3d <-
       warning("argument 'coords' is only used for 'mesh_plot(Raster)', ignoring")
     }
 
+
+    if (!is.null(x$material$texture)) {
+      x <- texture_mesh3d(x)
+    }
+
+    if (!is.null(crs)) {
+      xy <- try(reproj::reproj(t(x$vb[1:2, ]), crs)[,1:2], silent = TRUE)
+      if (!inherits(xy, "try-error")) x$vb[1:2, ] <- t(xy)
+    }
     if (!is.null(x$ib)) {
       id <- x$ib
     }
     if (!is.null(x$it)) {
       id <- x$it
     }
-    if (!is.null(x$material$texture)) {
-      warning("mesh object has a texture path, but cannot be displayed natively in 2D graphics (try plot3d)")
-      message("mesh plot will be displayed with an approximate colouring from the texture image")
-      ## ensure faces colour
-      x$meshColor <- "faces"
-      ## avoid rgdal lol so much pain avoided so easy
-      b <- raster::setExtent(raster::brick(
-        png::readPNG(x$material$texture)),
-                             raster::extent(0, 1, 0, 1))
-      rgb0 <- raster::extract(b, t(x$texcoords[1:2, ]))
-
-      red <- sqrt(colMeans(matrix(rgb0[id,   1] ^2, dim(id)[1L]), na.rm = TRUE))
-      green <- sqrt(colMeans(matrix(rgb0[id, 2] ^2, dim(id)[1L]), na.rm = TRUE))
-      blue <- sqrt(colMeans(matrix(rgb0[id,  3] ^2, dim(id)[1L]), na.rm = TRUE))
-#      x$material$color <- rgb(red, green, blue, maxColorValue = 255)
-      x$material$color <- colourvalues::convert_colour(cbind(red, green, blue) * 255)
-    }
-
     xx <- x$vb[1L, id]
     yy <- x$vb[2L, id]
 
@@ -222,6 +212,9 @@ mesh_plot.default <- function(x,  col = NULL, add = FALSE, zlim = NULL, ...,
 #' @export
 mesh_plot.triangulation <- function(x,  col = NULL, add = FALSE, zlim = NULL, ...,
                                     coords = NULL, crs = NULL) {
-  mesh_plot(as.mesh3d(x), col = col, add = add )
+  if (!is.null(coords)) {
+    warning("argument 'coords' is only used for 'mesh_plot(Raster)', ignoring")
+  }
+  mesh_plot(as.mesh3d(x), col = col, add = add, zlim = zlim, crs = crs)
 }